Created new clean RobotBuilder project and exported code.
Note* This uses the old layout of generating robot code that differes from the eclipse generated robot code. Major differences are no CommandBase (meaning objects are created in robot and robotMap but that shouldn't matter if we use robot builder) and RobotMap that holds the actuator objects intsead of constants for port numbers.